+++ tests/test_to_thread.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
 import asyncio
+import gc
+import sys
 import threading
 import time
 from concurrent.futures import Future, ThreadPoolExecutor
@@ -23,7 +25,7 @@ from anyio import (
 )
 from anyio.from_thread import BlockingPortalProvider
 
-from .conftest import asyncio_params
+from .conftest import asyncio_params, no_other_refs
 
 pytestmark = pytest.mark.anyio
 
@@ -360,3 +362,30 @@ class TestBlockingPortalProvider:
                 portal.call(event.set)
 
         assert len(threads) == 1
+
+
+@pytest.mark.skipif(
+    sys.implementation.name == "pypy",
+    reason=(
+        "gc.get_referrers is broken on PyPy (see "
+        "https://github.com/pypy/pypy/issues/5075)"
+    ),
+)
+async def test_run_sync_worker_cyclic_references() -> None:
+    class Foo:
+        pass
+
+    def foo(_: Foo) -> None:
+        pass
+
+    cvar = ContextVar[Foo]("cvar")
+    contextval = Foo()
+    arg = Foo()
+    cvar.set(contextval)
+    await to_thread.run_sync(foo, arg)
+    cvar.set(Foo())
+    gc.collect()
+
+    assert gc.get_referrers(contextval) == no_other_refs()
+    assert gc.get_referrers(foo) == no_other_refs()
+    assert gc.get_referrers(arg) == no_other_refs()